最終版本

+++ app/router/api/v1/prods/id/DELETE.go
@@ -4,6 +4,7 @@ import (
 	"foodlisa/models"
 	"foodlisa/router/baseHandler"
 	"github.com/gin-gonic/gin"
+	"strconv"
 )
 
 type DELETE struct {
@@ -26,11 +27,19 @@ func (obj *DELETE) Process() {
 		return
 	}
 
+	// 檢查是否是該用戶的商品
+	_user := obj.Ctx.MustGet("Id").(string)
+	loginUser, _ := strconv.Atoi(_user)
+	if loginUser != product.UserId {
+		obj.Ctx.Writer.WriteHeader(403)
+		return
+	}
+
 	// 刪除
 	result = obj.DB.Delete(&models.Product{}, obj.Ctx.Param("id"))
 
 	if result.Error == nil {
-		obj.Ctx.JSON(204, "")
+		obj.Ctx.Writer.WriteHeader(204)
 	} else {
 		obj.Ctx.JSON(400, gin.H{
 			"message": "Delete failed.",
